Add support for Renode.
To build for Renode we use the cc2538dk build target (TARGET), because
it's the only one supported by Contiki-NG.

To compile run:

    make RENODE=1

Setting RENODE=0 or not defining it defaults to a normal SimpleLink
CC1312R build.

Renode by default can't use the uart1 API, this limitation can't be
solved yet, a workaround needs to be made but now it's fine.

Signed-off-by: Jean Pierre Dudey <jeandudey@hotmail.com>