Request route to node using AODV.
This demonstration code requests a route to fe80::0:200:0:0:3,
you need to have three nodes in radio-nodes.resc script. A binary
named radio-firmware-recv.cc2538dk needs to be created that doesn't send any
RREQ and the normal binary radio-firmware.cc2538dk that sends a RREQ.

The RREQ travels to the requested address but it isn't able to send a
RREP back, more debug and testing needs to be done to check that.

You can test using

    make RENODE=1
    mv radio-firmware.cc2538dk radio-firmware-recv.cc2538dk

    # comment the line in radio-firmware.c that requests the route.
    # and compile again
    make RENODE=1

    # run renode, all you need to do is use the start command inside
    # Renode.
    renode radio-nodes.resc

Signed-off-by: Jean Pierre Dudey <jeandudey@hotmail.com>